test: make existing instance access portable to TT02
Tenor party IDs differ between localtest and TT02, and newly delegated
signing access is not reliable in TT02 due to authorization caching. Those
environment details obscured the frontend regression the test is meant to
protect.

Select the organization by its stable Tenor number and reopen the instance as
its original user. Mock only the instantiation-party response on the second
load so the test isolates the frontend guard while retaining real backend
instance authorization in both environments.
